dxf/export: move marshalSubtasks into scheduler, trim comments
marshalSubtasks does subtask-meta serialization and external-storage
offload, which is scheduler orchestration, so move it next to
OnNextSubtasksBatch and keep split.go focused on chunking. Drop the
now-unused proto/dxfutil/extstore imports from split.go.
